# artifacts/api-server/src/routes/notes.ts
import { Router, type IRouter } from "express";
import { eq, sql } from "drizzle-orm";
import { db, notesTable } from "@workspace/db";
import {
  CreateNoteBody,
  UpdateNoteBody,
  UpdateNoteParams,
  DeleteNoteParams,
  ToggleNoteParams,
} from "@workspace/api-zod";
import * as zod from "zod";

const router: IRouter = Router();

const CloneNoteParams = zod.object({ id: zod.coerce.number() });

// GET /notes
router.get("/notes", async (_req, res): Promise<void> => {
  const notes = await db
    .select()
    .from(notesTable)
    .orderBy(notesTable.createdAt);
  res.json(notes);
});

// GET /notes/stats
router.get("/notes/stats", async (_req, res): Promise<void> => {
  const [row] = await db
    .select({
      total: sql<number>`count(*)::int`,
      active: sql<number>`count(*) filter (where done = false)::int`,
      done: sql<number>`count(*) filter (where done = true)::int`,
    })
    .from(notesTable);
  res.json(row ?? { total: 0, active: 0, done: 0 });
});

// POST /notes
router.post("/notes", async (req, res): Promise<void> => {
  const parsed = CreateNoteBody.safeParse(req.body);
  if (!parsed.success) {
    res.status(400).json({ error: parsed.error.message });
    return;
  }

  const deadlineVal = (parsed.data as { deadline?: string | null }).deadline;
  const [note] = await db
    .insert(notesTable)
    .values({
      content: parsed.data.content,
      projectId: parsed.data.projectId ?? null,
      priority: parsed.data.priority ?? null,
      dayOfWeek: parsed.data.dayOfWeek ?? null,
      deadline: deadlineVal ? new Date(deadlineVal) : null,
    })
    .returning();

  res.status(201).json(note);
});

// POST /notes/:id/clone
router.post("/notes/:id/clone", async (req, res): Promise<void> => {
  const params = CloneNoteParams.safeParse(req.params);
  if (!params.success) {
    res.status(400).json({ error: params.error.message });
    return;
  }

  const [existing] = await db
    .select()
    .from(notesTable)
    .where(eq(notesTable.id, params.data.id));

  if (!existing) {
    res.status(404).json({ error: "Note not found" });
    return;
  }

  const [cloned] = await db
    .insert(notesTable)
    .values({
      content: `Копия: ${existing.content}`,
      projectId: existing.projectId,
      priority: existing.priority,
      dayOfWeek: existing.dayOfWeek,
      deadline: existing.deadline,
      done: false,
    })
    .returning();

  res.status(201).json(cloned);
});

// PATCH /notes/:id
router.patch("/notes/:id", async (req, res): Promise<void> => {
  const params = UpdateNoteParams.safeParse(req.params);
  if (!params.success) {
    res.status(400).json({ error: params.error.message });
    return;
  }

  const parsed = UpdateNoteBody.safeParse(req.body);
  if (!parsed.success) {
    res.status(400).json({ error: parsed.error.message });
    return;
  }

  const updateData: Record<string, unknown> = {};
  if (parsed.data.content !== undefined) updateData.content = parsed.data.content;
  if ("projectId" in parsed.data) updateData.projectId = parsed.data.projectId ?? null;
  if ("priority" in parsed.data) updateData.priority = parsed.data.priority ?? null;
  if ("dayOfWeek" in parsed.data) updateData.dayOfWeek = parsed.data.dayOfWeek ?? null;

  // Handle deadline field
  const deadlineVal = (parsed.data as { deadline?: string | null | undefined }).deadline;
  if ("deadline" in parsed.data) {
    updateData.deadline = deadlineVal ? new Date(deadlineVal) : null;
  }

  const [note] = await db
    .update(notesTable)
    .set(updateData)
    .where(eq(notesTable.id, params.data.id))
    .returning();

  if (!note) {
    res.status(404).json({ error: "Note not found" });
    return;
  }

  res.json(note);
});

// DELETE /notes/:id
router.delete("/notes/:id", async (req, res): Promise<void> => {
  const params = DeleteNoteParams.safeParse(req.params);
  if (!params.success) {
    res.status(400).json({ error: params.error.message });
    return;
  }

  const [note] = await db
    .delete(notesTable)
    .where(eq(notesTable.id, params.data.id))
    .returning();

  if (!note) {
    res.status(404).json({ error: "Note not found" });
    return;
  }

  res.sendStatus(204);
});

// PATCH /notes/:id/toggle
router.patch("/notes/:id/toggle", async (req, res): Promise<void> => {
  const params = ToggleNoteParams.safeParse(req.params);
  if (!params.success) {
    res.status(400).json({ error: params.error.message });
    return;
  }

  const [existing] = await db
    .select()
    .from(notesTable)
    .where(eq(notesTable.id, params.data.id));

  if (!existing) {
    res.status(404).json({ error: "Note not found" });
    return;
  }

  const [note] = await db
    .update(notesTable)
    .set({ done: !existing.done })
    .where(eq(notesTable.id, params.data.id))
    .returning();

  res.json(note);
});

export default router;