# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Project {
  id: number;
  name: string;
  color: string;
  createdAt: string;
}

export interface ProjectInput {
  /** @minLength 1 */
  name: string;
  /** @minLength 1 */
  color: string;
}

export interface ProjectUpdate {
  /** @minLength 1 */
  name?: string;
  /** @minLength 1 */
  color?: string;
}

export interface Note {
  id: number;
  content: string;
  done: boolean;
  /** @nullable */
  projectId: number | null;
  /** @nullable */
  priority: 'high' | 'medium' | 'low' | null;
  /** 0=Mon … 6=Sun @nullable */
  dayOfWeek: number | null;
  /** Optional deadline timestamp @nullable */
  deadline: string | null;
  createdAt: string;
  updatedAt: string;
}

export interface NoteInput {
  /** @minLength 1 */
  content: string;
  /** @nullable */
  projectId?: number | null;
  /** @nullable */
  priority?: 'high' | 'medium' | 'low' | null;
  /** 0=Mon … 6=Sun @nullable */
  dayOfWeek?: number | null;
  /** Optional deadline ISO string @nullable */
  deadline?: string | null;
}

export interface NoteUpdate {
  /** @minLength 1 */
  content?: string;
  /** @nullable */
  projectId?: number | null;
  /** @nullable */
  priority?: 'high' | 'medium' | 'low' | null;
  /** 0=Mon … 6=Sun @nullable */
  dayOfWeek?: number | null;
  /** Optional deadline ISO string @nullable */
  deadline?: string | null;
}

export interface NoteStats {
  total: number;
  active: number;
  done: number;
}

export interface VaultEntry {
  id: number;
  title: string;
  content: string;
  createdAt: string;
  updatedAt: string;
}

export interface VaultEntryInput {
  /** @minLength 1 */
  title: string;
  /** @minLength 1 */
  content: string;
}

export interface VaultEntryUpdate {
  /** @minLength 1 */
  title?: string;
  /** @minLength 1 */
  content?: string;
}

export interface ApiError {
  error: string;
}