# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: notes
    description: Notes operations
  - name: projects
    description: Projects operations
  - name: vault
    description: Vault operations
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /projects:
    get:
      operationId: listProjects
      tags: [projects]
      summary: List all projects
      responses:
        "200":
          description: List of projects
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Project"
    post:
      operationId: createProject
      tags: [projects]
      summary: Create a project
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProjectInput"
      responses:
        "201":
          description: Created project
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"

  /projects/{id}:
    patch:
      operationId: updateProject
      tags: [projects]
      summary: Update a project
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProjectUpdate"
      responses:
        "200":
          description: Updated project
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
    delete:
      operationId: deleteProject
      tags: [projects]
      summary: Delete a project
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"

  /notes:
    get:
      operationId: listNotes
      tags: [notes]
      summary: List all notes
      responses:
        "200":
          description: List of notes
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Note"
    post:
      operationId: createNote
      tags: [notes]
      summary: Create a note
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NoteInput"
      responses:
        "201":
          description: Created note
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Note"

  /notes/stats:
    get:
      operationId: getNoteStats
      tags: [notes]
      summary: Get note statistics
      responses:
        "200":
          description: Note statistics
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NoteStats"

  /notes/{id}:
    patch:
      operationId: updateNote
      tags: [notes]
      summary: Update a note
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NoteUpdate"
      responses:
        "200":
          description: Updated note
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Note"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
    delete:
      operationId: deleteNote
      tags: [notes]
      summary: Delete a note
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"

  /notes/{id}/toggle:
    patch:
      operationId: toggleNote
      tags: [notes]
      summary: Toggle done status of a note
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Updated note
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Note"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"

  /notes/{id}/clone:
    post:
      operationId: cloneNote
      tags: [notes]
      summary: Clone a note
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "201":
          description: Cloned note
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Note"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"

  /vault:
    get:
      operationId: listVaultEntries
      tags: [vault]
      summary: List all vault entries
      responses:
        "200":
          description: List of vault entries
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/VaultEntry"
    post:
      operationId: createVaultEntry
      tags: [vault]
      summary: Create a vault entry
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VaultEntryInput"
      responses:
        "201":
          description: Created vault entry
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VaultEntry"

  /vault/{id}:
    patch:
      operationId: updateVaultEntry
      tags: [vault]
      summary: Update a vault entry
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VaultEntryUpdate"
      responses:
        "200":
          description: Updated vault entry
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VaultEntry"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
    delete:
      operationId: deleteVaultEntry
      tags: [vault]
      summary: Delete a vault entry
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    Project:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        color:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - color
        - createdAt

    ProjectInput:
      type: object
      properties:
        name:
          type: string
          minLength: 1
        color:
          type: string
          minLength: 1
      required:
        - name
        - color

    ProjectUpdate:
      type: object
      properties:
        name:
          type: string
          minLength: 1
        color:
          type: string
          minLength: 1

    Note:
      type: object
      properties:
        id:
          type: integer
        content:
          type: string
        done:
          type: boolean
        projectId:
          type: ["integer", "null"]
        priority:
          type: ["string", "null"]
          enum: ["high", "medium", "low", null]
          description: "high | medium | low | null"
          nullable: true
        dayOfWeek:
          type: ["integer", "null"]
          description: "0=Mon … 6=Sun"
          nullable: true
        deadline:
          type: ["string", "null"]
          format: date-time
          description: "Optional deadline timestamp"
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - content
        - done
        - projectId
        - createdAt
        - updatedAt

    NoteInput:
      type: object
      properties:
        content:
          type: string
          minLength: 1
        projectId:
          type: ["integer", "null"]
        priority:
          type: ["string", "null"]
          enum: ["high", "medium", "low", null]
          nullable: true
        dayOfWeek:
          type: ["integer", "null"]
          description: "0=Mon … 6=Sun"
          nullable: true
        deadline:
          type: ["string", "null"]
          format: date-time
          nullable: true
      required:
        - content

    NoteUpdate:
      type: object
      properties:
        content:
          type: string
          minLength: 1
        projectId:
          type: ["integer", "null"]
        priority:
          type: ["string", "null"]
          enum: ["high", "medium", "low", null]
          nullable: true
        dayOfWeek:
          type: ["integer", "null"]
          nullable: true
        deadline:
          type: ["string", "null"]
          format: date-time
          nullable: true

    NoteStats:
      type: object
      properties:
        total:
          type: integer
        active:
          type: integer
        done:
          type: integer
      required:
        - total
        - active
        - done

    VaultEntry:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        content:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - title
        - content
        - createdAt
        - updatedAt

    VaultEntryInput:
      type: object
      properties:
        title:
          type: string
          minLength: 1
        content:
          type: string
          minLength: 1
      required:
        - title
        - content

    VaultEntryUpdate:
      type: object
      properties:
        title:
          type: string
          minLength: 1
        content:
          type: string
          minLength: 1

    ApiError:
      type: object
      properties:
        error:
          type: string
      required:
        - error